Move the custom property expansion code to a separate file
