+++ lib/util/fonts/getTextByFontProperties.js
@@ -3,7 +3,6 @@ const defaultStylesheets = require('./defaultStylesheets');
 const stylePropObjectComparator = require('./stylePropObjectComparator');
 const unquote = require('./unquote');
 const memoizeSync = require('memoizesync');
-const postCssValuesParser = require('postcss-values-parser');
 const cssFontWeightNames = require('css-font-weight-names');
 const capitalize = require('capitalize');
 const cssPseudoElementRegExp = require('../cssPseudoElementRegExp');
@@ -17,6 +16,8 @@ const getCounterCharacters = require('./getCounterCharacters');
 const expandPermutations = require('../expandPermutations');
 const combinePredicates = require('./combinePredicates');
 const arePredicatesExhaustive = require('../arePredicatesExhaustive');
+const initialValueByProp = require('./initialValueByProp');
+const expandCustomProperties = require('./expandCustomProperties');
 
 const FONT_PROPS = ['font-family', 'font-style', 'font-weight'];
 
@@ -37,24 +38,6 @@ const CSS_PROPS_TO_TRACE = [
 ];
 
 const CSS_PROPS_TO_TRACE_AND_TEXT = ['text', ...CSS_PROPS_TO_TRACE];
-
-const INITIAL_VALUES = {
-  // 'font-family': 'serif'
-  'font-weight': 400,
-  'font-style': 'normal',
-  content: 'normal',
-  quotes: '"«" "»" "‹" "›" "‘" "’" "\'" "\'" "\\"" "\\""', // Wide default set to account for browser differences
-  'list-style-type': 'none',
-  display: 'inline',
-  'animation-name': 'none',
-  'text-transform': 'none',
-  'transition-property': 'all',
-  'transition-duration': '0s',
-  'counter-increment': 'none',
-  'counter-reset': 'none',
-  'counter-set': 'none',
-  'white-space': 'normal'
-};
 
 const INHERITED = {
   'font-family': true,
@@ -181,7 +164,7 @@ function getMemoizedElementStyleResolver(
       // Stop condition. We moved above <HTML>
       if (!node.tagName) {
         cssPropsAndCustomPropsToTrace.forEach(prop => {
-          result[prop] = [{ value: INITIAL_VALUES[prop], predicates }];
+          result[prop] = [{ value: initialValueByProp[prop], predicates }];
         });
         return result;
       }
@@ -304,7 +287,7 @@ function getMemoizedElementStyleResolver(
             } else {
               let value;
               if (declaration.value === 'initial') {
-                value = INITIAL_VALUES[prop];
+                value = initialValueByProp[prop];
               } else if (
                 NAME_CONVERSIONS[prop] &&
                 NAME_CONVERSIONS[prop][declaration.value]
@@ -369,103 +352,7 @@ function getMemoizedElementStyleResolver(
               });
               hypotheticalValues = multipliedHypotheticalValues;
             }
-            const hypotheticalValuesByCustomProp = {};
-            for (let i = 0; i < hypotheticalValues.length; i += 1) {
-              const hypotheticalValue = hypotheticalValues[i];
-              // Quick test for whether the value contains custom properties:
-              if (/var\(--[^)]+\)/.test(hypotheticalValue.value)) {
-                const tokens = postCssValuesParser(hypotheticalValue.value)
-                  .tokens;
-                const seenCustomProperties = new Set();
-                for (let j = 0; j < tokens.length - 3; j += 1) {
-                  if (
-                    tokens[j][1] === 'var' &&
-                    tokens[j + 1][0] === '(' &&
-                    tokens[j + 2][1] === '--' &&
-                    tokens[j + 3][0] === 'word'
-                  ) {
-                    const customPropertyName = '--' + tokens[j + 3][1];
-                    hypotheticalValuesByCustomProp[
-                      customPropertyName
-                    ] = hypotheticalValuesByCustomProp[customPropertyName] ||
-                      traceProp(
-                        customPropertyName,
-                        0,
-                        hypotheticalValue.predicates
-                      ) || [
-                        {
-                          value: undefined,
-                          predicates: hypotheticalValue.predicates
-                        }
-                      ];
-                    seenCustomProperties.add(customPropertyName);
-                  }
-                }
-                if (seenCustomProperties.size === 0) {
-                  // The quick regexp test was a false positive
-                  continue;
-                }
-                const replacementHypotheticalValues = [];
-                for (const permutation of expandPermutations(
-                  hypotheticalValuesByCustomProp
-                )) {
-                  const predicates = combinePredicates([
-                    hypotheticalValue.predicates,
-                    ...Object.values(permutation).map(v => v.predicates)
-                  ]);
-                  let value = '';
-                  for (let j = 0; j < tokens.length; j += 1) {
-                    if (
-                      tokens[j][1] === 'var' &&
-                      tokens[j + 1][0] === '(' &&
-                      tokens[j + 2][1] === '--' &&
-                      tokens[j + 3][0] === 'word'
-                    ) {
-                      const customPropertyName = `--${tokens[j + 3][1]}`;
-                      if (
-                        permutation[customPropertyName] &&
-                        permutation[customPropertyName].value
-                      ) {
-                        value += permutation[customPropertyName].value;
-                        while (j < tokens.length && tokens[j][0] !== ')') {
-                          j += 1;
-                        }
-                      } else if (
-                        j < tokens.length &&
-                        tokens[j + 4][0] === 'comma'
-                      ) {
-                        // Undefined property, but there is a default value
-                        j += 5;
-                        while (j < tokens.length && tokens[j][0] === 'space') {
-                          j += 1;
-                        }
-                        while (j < tokens.length && tokens[j][0] !== ')') {
-                          value += tokens[j][1];
-                          j += 1;
-                        }
-                      } else {
-                        // Reference to an undefined custom property and no default value
-                        value = INITIAL_VALUES[prop];
-                        break;
-                      }
-                    } else {
-                      value += tokens[j][1];
-                    }
-                  }
-                  replacementHypotheticalValues.push({
-                    predicates,
-                    value
-                  });
-                }
-                hypotheticalValues.splice(
-                  i,
-                  1,
-                  ...replacementHypotheticalValues
-                );
-                i -= 1;
-              }
-            }
-            return hypotheticalValues;
+            return expandCustomProperties(hypotheticalValues, prop, traceProp);
           }
         }
         if (!nonInheritingTags.includes(node.tagName)) {
@@ -478,7 +365,7 @@ function getMemoizedElementStyleResolver(
               predicates
             ))[prop];
         } else {
-          return [{ value: INITIAL_VALUES[prop], predicates }];
+          return [{ value: initialValueByProp[prop], predicates }];
         }
       }
