+++ lib/util/fonts/initialValueByProp.js
@@ -1,0 +1,17 @@
+module.exports = {
+  // 'font-family': 'serif'
+  'font-weight': 400,
+  'font-style': 'normal',
+  content: 'normal',
+  quotes: '"«" "»" "‹" "›" "‘" "’" "\'" "\'" "\\"" "\\""', // Wide default set to account for browser differences
+  'list-style-type': 'none',
+  display: 'inline',
+  'animation-name': 'none',
+  'text-transform': 'none',
+  'transition-property': 'all',
+  'transition-duration': '0s',
+  'counter-increment': 'none',
+  'counter-reset': 'none',
+  'counter-set': 'none',
+  'white-space': 'normal'
+};
